mconline/requirements.txt
Flask
requests
python-dotenv
Werkzeug
Jinja2